docker-manager-exec/requirements.txt
docker==4.1.0
Flask==1.0.2
Flask-Sockets==0.2.1
PyYAML==5.1.2